Add support for GIC
According to the specification, a virtual Generic Interrupt Controller
(vGIC) should be presented to a Realm in order for the Realm to use
interrupts. The vGIC is implemented via a combination of Host emulation
and RMM mediation.

For the RMM mediation, this commit restores and saves the interrupt
context on REC entry and exit by communicating with the Host.