rmm: Add support for MMIO handling
This commit helps to handle the Memory-mapped I/O (MMIO) accesses by Realms.
When a Realm accesses device's registers using MMIO, it causes a data abort
which was the main reason for the kernel stuck. In this commit, RMM receives
the emulated value from the Host (nw-linux) and injects the value to the
Realm after setting the Realm's PC value to the next point. With these fixes,
the guest kernel in the Realm now could boot to the next stage.

Reference: Arm Armv9-A Architecture Registers
           TF-RMM (https://github.com/TF-RMM/tf-rmm)

[before]
...
[    0.000000] rcu: Adjusting geometry for rcu_fanout_leaf=16, nr_cpu_ids=1
[    0.000000] NR_IRQS: 64, nr_irqs: 64, preallocated irqs: 0
       //  no more log after this point  //

[after]
...
[    0.000000] rcu: Adjusting geometry for rcu_fanout_leaf=16, nr_cpu_ids=1
[    0.000000] NR_IRQS: 64, nr_irqs: 64, preallocated irqs: 0
[    0.000000] GICv3: 32 SPIs implemented
[    0.000000] GICv3: 0 Extended SPIs implemented
[    0.000000] Root IRQ handler: gic_handle_irq
[    0.000000] GICv3: GICv3 features: 16 PPIs, DirectLPI
[    0.000000] GICv3: CPU0: found redistributor 0 region 0:0x000000003ffd0000
[    0.000000] rcu: srcu_init: Setting srcu_struct sizes based on contention.
[    0.000000] arch_timer: cp15 timer(s) running at 100.00MHz (virt).
...
[    0.692188] RPC: Registered tcp NFSv4.1 backchannel transport module.
[    0.702813] kvm [1]: HYP mode not available
[    0.710786] Initialise system trusted keyrings
[    0.718649] Unpacking initramfs...