rmm: Add a level to page table translation
In page table traversing API, (input) level and (output) level are
added which is useful for the fine-grained traversing. (input) level
indicates the intended (maximum) level to reach, while (output) level
is the actual level reached. For example, if the user passes 3 for
(input) level and there are upto level 2 page tables for the given
address, it would return 2 for the (output) level.