sdk: Fix simulated verify condition for user data
[Background]
When verify() is called on x64, which means simulated version,
SDK replaced hard_coded user_date(Realm Challenge) to given user_data.
Given user_data came from attest().

[Problem]
When an attestation report is generated on aarch64
`user_data` of attest() is encoded to the Realm Challenge Claim.

In cross-platform-e2ee scenario,
the Certifier Service verifies the report.

Since the Certifier Service runs on x64,
SDK replaced the Realm Challenge Claim to the given user data.

But the report is generated on x64, the given user data is not exist.

[This patch]
Do not replace the Realm Challenge Claim when the user data is empty.

Signed-off-by: Sangwan Kwon <sangwan.kwon@samsung.com>