Update RSI kernel module for EAC5 standard
It uses new RSI API for VERSION, ATTESTATION_INIT and ATTESTATION_CONTINUE
calls. Due to the changes in those calls it also updates the ioctl API it
provides as the output token can be longer than 4k now. The ioctl API handles
such cases. It's usage is described briefly in rsi.h header.

Rust wrapper for the ioctl API is updated in the remote-attestation repository
on eac5 branc.