refactor: Get rmi constraints without insertion
This commit refactors RMI's constraint retrieval procedure. The main
motivation of the refactoring is that insertion & get in the map are
very expensive operations especially in the formal semantic and avoiding
them is better if possible. It would be also beneficial for runtime
performance, as the previous insertions are no longer involved. The
assumption is that the configuration regarding constraints does not
change at runtime.

Signed-off-by: Changho Choi <ch754.choi@samsung.com>