mc: Add host memory modeling code
The Host delegates and undelegates its memory for Realm World's
usage. To verify RMM, the host memory region should be properly
modeled.

We model the Host memory as a pre-allocated memory region which
can avoid a false positive related to invalid memory accesses
in model checking. Also, instead of using the same starting
address (e.g., 0x8000_0000), we use a mock region filled with
non-deterministic contents. It helps to address an issue related
to the backend CBMC's pointer encoding, as 0x8000_0000 cannot be
distinguished from null pointer in CBMC.

Signed-off-by: Changho Choi <ch754.choi@samsung.com>