refactor: Do not import the whole rmm into uart
Importing the whole rmm is not a good practice for the library (uart).
The change only imports the necessary parts of rmm (`io` and `error`).
For this purpose, `io` and `error` are relocated as a library instead
of the previous rmm's modules.

```
[before]
└── uart v0.0.1 (/home/changho/zpzigi754/islet/lib/uart)
    ├── islet_rmm v0.0.1 (/home/changho/zpzigi754/islet/rmm) (*)
    ...

[after]
uart v0.0.1 (/home/changho/zpzigi754/islet/lib/uart)
├── io v0.0.1 (/home/changho/zpzigi754/islet/lib/io)
...
```

Signed-off-by: Changho Choi <ch754.choi@samsung.com>