rmm: Avoid double copy on RMI_DATA_CREATE
To mitigate concurrency issues with memory regions shared
between the host and RMM, the host::copy_from operation was
used to copy data from the host.

However, for the RMI_DATA_CREATE case, this mitigation
let to double copy costs.

To address this, a new function host::copy_to_obj has been
added to perform the copy in a single step.

Signed-off-by: Sangwan Kwon <sangwan.kwon@samsung.com>