spec: feedback: Ensuring realm_state validation in B4.3.15
It seems necessary to review between B4.3.15 and D1.2.2 for spec consistency.

1. Specification Restrictions
According to "D1.2.2 Realm Translation Table creation flow" subsequent levels of RTT
can be created when the state of the Realm is REALM_NEW or REALM_ACTIVATE.

> D1.2.2 Realm Translation Table creation flow
> Subsequent levels of RTT are added using the RMI_RTT_CREATE command.
> This can be performed when the state of the Realm is REALM_NEW or REALM_ACTIVE.

2. Missing Failure Conditions
However in "B4.3.15 RMI_RTT_CREATE", the condition to check the realm_state is missing.

3. Suggestion
It would be necessary either to add this realm_state validation within "B4.3.15.2 Failure conditions"
or to clarify the details in D1.2.2 for better alignment.

Signed-off-by: Sangwan Kwon <sangwan.kwon@samsung.com>