scripts: Set acs's EXPECTED to 74
This commit temporarily reduces EXPECTED by 2 to pass CI tests
(mm_rtt_level_start and mm_ha_hd_access).
During the test of `mm_rtt_level_start`, a granule failed to be
undelegated, because it was in the state of RTT: this means that
RTT granule has not been destroyed well at a certain point because
of another issue. I will resolve this issue in another commit.

```
[before]
[TRACE]islet_rmm::event -- RMI: GRANULE_DELEGATE     [883C5000] > [0]    <-- delegated here
...
[TRACE]islet_rmm::event -- RMI: GRANULE_UNDELEGATE   [887E6000] > [0]
[TRACE]islet_rmm::event -- RMI: REALM_DESTROY        [88781000] > [0]
[TRACE]islet_rmm::event -- RMI: GRANULE_UNDELEGATE   [883C4000] > [0]
[TRACE]islet_rmm::event -- RMI: GRANULE_UNDELEGATE   [883C5000] > [0]    <-- undelegated well
```

```
[after]
[TRACE]islet_rmm::event -- RMI: GRANULE_DELEGATE     [883C5000] > [0]    <-- delegated here
...
[TRACE]islet_rmm::event -- RMI: GRANULE_UNDELEGATE   [887E6000] > [0]
[TRACE]islet_rmm::event -- RMI: REALM_DESTROY        [88781000] > [0]
[TRACE]islet_rmm::event -- RMI: GRANULE_UNDELEGATE   [883C4000] > [0]
[TRACE]islet_rmm::event -- RMI: GRANULE_UNDELEGATE   [883C5000] > [1]    <-- undelegation failed
```

Signed-off-by: Changho Choi <ch754.choi@samsung.com>