spec: feedback: Remove a failure condition
The below feedback would be accepted in 1.1-alp10.

[Feedback]
8466a917 spec: feedback: Ensuring realm_state validation in B4.3.1

[Reply]
The flow description has been amended, to remove the constraint on the Realm state.
This change will be included in 1.1-alp10.

that is: ( the last line is removed)
    D1.2.2 Realm Translation Table creation flow
    Subsequent levels of RTT are added using the RMI_RTT_CREATE command.
(-) This can be performed when the state of the Realm is REALM_NEW or REALM_ACTIVE.

Signed-off-by: Sangwan Kwon <sangwan.kwon@samsung.com>