Update TF-RMM and TF-A (v2.11)
- TF-RMM (fdcf08)
- TF-A (v2.11)

This commit prepares for using 1.0-REL0.

TF-RMM's rmm-spec-v1.0-rel0 branch has a compatibility issue,
so this commit updates TF-RMM and TF-A which do not cause that issue.

fdcf08 is the last commit before the commit breaking a backward
compatibility (715ff).

According to 715ff, it says that RMM would expect RMM-EL3 interface
v0.3, but there seems no build option for enforcing this version
of interface, which causes almost all ACS tests to fail.

Until this issue is resolved, it would be better to use an older
version (an incomplete 1.0-REL0).

Signed-off-by: Changho Choi <ch754.choi@samsung.com>