Do not use duplicate bit positions for RIPAS and HIPAS
Bit position 6 of S2TTE is aliasing for RIPAS and S2AP in TF-RMM
and previous Islet. This aliasing causes a confusion and issue.
For example, after the execution of RMI_DATA_CREATE,
RIPAS field was set as DESTROYED while it should be set as RAM.
TF-RMM circumvents this issue by using a certain signature in S2TTE
(e.g., interpret DESTROYED page as RAM page with that signature).

We instead allocate and use different positions for RIPAS and HIPAS
to avoid such an aliasing and confusion. This helps to pass relevant
ACS tests without using the signature.

Signed-off-by: Changho Choi <ch754.choi@samsung.com>