Update rmi and rsi constraints
The previous constraints did not reflect the spec well (wrong or missing
values). This patch resolve this issue in part.

Note that RSI and PSCI constraint values are currently not enforced, as
arg and ret values in Context are not used in the handlers where
set_reg and get_reg are instead used.

Signed-off-by: Changho Choi <ch754.choi@samsung.com>