Modify HES TF-A build options
MEASURED_BOOT cannot be used at the time with TF-A. As usual it doesn't fit in
BL31 proglimits and link time optimizations we used to use cause the whole setup
to become unstable.

For now use HES just with delegated attestation, the token is still properly
prepared, just without bootloader measurements in platform token.

Signed-off-by: Lukasz Pawelczyk <l.pawelczyk@samsung.com>