Update confidential-ml example to be a standalone example
This commit gets confidential-ml example to be a standalone example,
which you can try from building from sources to running them on the host machine
without relying on a special docker image.

This commit only supports simulated-enclave for now, meaning that
running "device" binary on realm is not supported.
This realm-execution supports will come in a next PR.

Signed-off-by: Jinbum Park <jinb.park@samsung.com>